+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Implement `IBucket.CreateQueryLink` method, [PR-93](https://github.com/reductstore/reduct-cpp/pull/93)
+
 ## [1.16.0] - 2025-08-12
 
 ## Added